+++ src/lib/C4Rect.cpp
@@ -177,7 +177,7 @@ void C4Rect::Add(const C4Rect &r2)
 void C4RectList::ClipByRect(const C4Rect &rClip)
 {
 	// split up all rectangles
-	for (int32_t i = 0; i < GetCount(); ++i)
+	for (size_t i = 0; i < GetCount(); ++i)
 	{
 		C4Rect *pTarget = &Get(i);
 		// any overlap?
@@ -228,10 +228,10 @@ void C4RectList::ClipByRect(const C4Rect &rClip)
 	while (!fDone)
 	{
 		fDone=true;
-		for (int32_t i = 0, cnt=GetCount(); i < cnt && fDone; ++i)
+		for (size_t i = 0, cnt=GetCount(); i < cnt && fDone; ++i)
 		{
 			C4Rect &rc1 = Get(i);
-			for (int32_t j = i+1; j < cnt; ++j)
+			for (size_t j = i+1; j < cnt; ++j)
 			{
 				C4Rect &rc2 = Get(j);
 				if (rc1.y == rc2.y && rc1.Hgt == rc2.Hgt)
