+++ src/lib/C4Rect.h
@@ -105,7 +105,7 @@ public:
 	void AddRect(const C4Rect &rNewRect)
 	{ push_back(rNewRect); }
 	void RemoveIndexedRect(int32_t idx)
-	{ if (idx<GetCount()-1) Get(idx)=Get(GetCount()-1); pop_back(); }
+	{ if (idx<int32_t(GetCount()-1)) Get(idx)=Get(GetCount()-1); pop_back(); }
 	void Clear() { clear(); }
 	size_t GetCount() const { return size(); }
 	C4Rect &Get(size_t idx) { return (*this)[idx]; } // access w/o range check
